Train a CNN on the SVHN Dataset for Classification
The Street View House Numbers (SVHN) dataset is a challenging real-world dataset consisting of images of house numbers captured in urban environments. The main goal of this project is to build a robust machine learning model capable of classifying digits from these images. SVHN presents a more difficult task compared to MNIST due to varying lighting conditions, distortions, and overlapping digits.

Convolutional Neural Networks (CNNs) are the state-of-the-art approach for image classification tasks due to their ability to automatically learn spatial hierarchies of features. This project leverages CNNs to address the classification challenge of the SVHN dataset, aiming to build a model that can accurately classify digits captured in various street view conditions.